refactor: move more RequiredVersion tests from utils to domain test files
